requirements.txt
asgiref==3.6.0
Django==4.1.5
django-appconf==1.0.5
django-compressor==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
rcssmin==1.1.1
rjsmin==1.2.1
sqlparse==0.4.3
whitenoise==6.3.0
